+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Removed `MockDataStore` and replaced usage in affected tests for loading mock
   data on the DB.
 * Removed the `mock-data` command from the CLI
+* Removed `MockClient` altogether from the CLI
 
 ## 0.1.0 (2024-03-15)